Externalise logic for comparing Registration rerlated objects
From https://github.com/DEFRA/ruby-services-team/issues/44

Create a service to compare registration and edit registration related objects and decide if the registration has been edited or not.
