+++ app/services/waste_exemptions_engine/edit_completion_service.rb
@@ -19,8 +19,6 @@ module WasteExemptionsEngine
     end
 
     def copy_data_from_edit_registration
-      related_objects_changed?
-
       copy_attributes
       copy_addresses
       copy_people
@@ -55,31 +53,9 @@ module WasteExemptionsEngine
     def save_registration_if_changed
       if @registration.changed?
         @registration.save!
-      elsif related_objects_changed?
+      elsif @registration.related_objects_changed?(@edit_registration)
         @registration.paper_trail.save_with_version
       end
-    end
-
-    def related_objects_changed?
-      @related_objects_changed ||= address_data_changed? || people_data_changed?
-    end
-
-    def address_data_changed?
-      old_data = comparable_data(@registration.addresses.order(:address_type))
-      new_data = comparable_data(@edit_registration.addresses.order(:address_type))
-
-      old_data != new_data
-    end
-
-    def people_data_changed?
-      old_data = comparable_data(@registration.people.order(:first_name))
-      new_data = comparable_data(@edit_registration.people.order(:first_name))
-
-      old_data != new_data
-    end
-
-    def comparable_data(items)
-      items.to_json(except: %w[id created_at updated_at registration_id transient_registration_id])
     end
   end
 end
